fix: runtime error index out of range caused by nextCleanupBucket
- Adjusted cache size and TTL in `sieve_test.go` for more comprehensive testing
- Implemented a constant `numberOfBuckets` and utilized it in `sieve.go` for cache storage
- Updated `addToBucket` function to incorporate the new `numberOfBuckets` constant and `int8` `bucketID` type
- Modified `bucketId` calculation in `addToBucket` function using `s.nextCleanupBucket` of type `float64`
- Enhanced workload for testing in `sieve_test.go` from 256 to 10240 for rigorous testing

+++ sieve/sieve.go
@@ -9,6 +9,10 @@ import (
 	"github.com/scalalang2/golang-fifo/types"
 )
 
+// numberOfBuckets is the number of buckets to store the cache entries
+//
+// Notice: if this number exceeds 256, the type of nextCleanupBucket
+// in the Sieve struct should be changed to int16
 const numberOfBuckets = 100
 
 // entry holds the key and value of a cache entry.
@@ -249,8 +253,8 @@ func (s *Sieve[K, V]) addToBucket(e *entry[K, V]) {
 	if s.ttl == 0 {
 		return
 	}
-	bucketId := (numberOfBuckets + s.nextCleanupBucket - 1) % numberOfBuckets
-	e.bucketID = bucketId
+	bucketId := (numberOfBuckets + int(s.nextCleanupBucket) - 1) % numberOfBuckets
+	e.bucketID = int8(bucketId)
 	s.buckets[bucketId].entries[e.key] = e
 	if s.buckets[bucketId].newestEntry.Before(e.expiredAt) {
 		s.buckets[bucketId].newestEntry = e.expiredAt
